deprecating tools.synthetic_pii, and moving all but DOB into components.synthetic_pii
